Atualização das classes - Cliente

+++ Repository/ClienteRepository.cs
@@ -1,14 +1,11 @@
 ﻿using Domain;
 using Infradb;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Microsoft.EntityFrameworkCore;
+using Repository.Interfaces;
 
 namespace Repository
 {
-    public class ClienteRepository
+    public class ClienteRepository : IClienteRepository
     {
         private readonly AppDbContext _context;
 
@@ -24,9 +21,9 @@ namespace Repository
             return cliente;
         }
 
-        public async Task<List<Cliente>> ObterClientes()
+        public async Task<List<Cliente>> ObterClientes(int usuarioId)
         {
-            return await _context.Clientes.ToListAsync();
+            return await _context.Clientes.Where(c => c.Id_usuario == usuarioId).ToListAsync();
         }
 
         public async Task<Cliente?> ObterClienteId(int id)
@@ -37,13 +34,13 @@ namespace Repository
         public async Task<Cliente?> AtualizarCliente(int id, Cliente clienteAtualizado)
         {
             var cliente = await _context.Clientes.FindAsync(id);
-            if(cliente == null) return null;
+            if(cliente == null || cliente.Id_usuario != clienteAtualizado.Id_usuario) return null;
 
-            cliente.Nome_cliente = clienteAtualizado.NomeCliente;
+            cliente.Nome_cliente = clienteAtualizado.Nome_cliente;
             cliente.Telefone = clienteAtualizado.Telefone;
-            cliente.Email_cliente = clienteAtualizado.EmailCliente;
-            cliente.Endereco_rua = clienteAtualizado.EnderecoRua;
-            cliente.Numero_residencia = clienteAtualizado.NumeroResidencia;
+            cliente.Email_cliente = clienteAtualizado.Email_cliente;
+            cliente.Endereco_rua = clienteAtualizado.Endereco_rua;
+            cliente.Numero_residencia = clienteAtualizado.Numero_residencia;
             cliente.Bairro = clienteAtualizado.Bairro;
             cliente.Cidade = clienteAtualizado.Cidade;
             cliente.Estado = clienteAtualizado.Estado;
@@ -52,10 +49,10 @@ namespace Repository
             return cliente;
         }
 
-        public async Task<bool> DeletarCliente(int id)
+        public async Task<bool> DeletarCliente(int id, int usuarioId)
         {
-            var cliente = await _context.Clientes.FindAsync(id);
-            if(cliente == null) return false;
+            var cliente = await _context.Clientes.FirstOrDefaultAsync(s => s.Id_cliente == id && s.Id_usuario == usuarioId);
+            if(cliente == null || cliente.Id_usuario != usuarioId) return false;
 
             _context.Clientes.Remove(cliente);
             await _context.SaveChangesAsync();
